refactor: share one implementation between the public and private pin routes
`pins.tsx` and `private.tsx` were near-identical copies -- 558 lines byte-for-byte
between them, handler for handler. Both now mount `createPinRoutes` from the new
`pin-routes.tsx`, configured by `baseUrl` and `privateMode`.

    pins.tsx     739 ->  10
    private.tsx  689 ->  85   (unlock/lock, which has no public counterpart)
    pin-routes.tsx    -> 703
    total       1428 -> 798

Everything the two genuinely differ on is now an explicit option rather than a
divergence between two files that had to be kept in sync by hand:

- which links get generated (baseUrl)
- whether the list filters to private pins
- whether a new pin is forced private regardless of the submitted value
- whether the pages render the private chrome
- the create flash wording

Two subtleties the copies encoded implicitly, now written down at the point they
matter:

- create forces isPrivate, edit honours the submitted value. Preserved and
  commented; unifying them would silently change whether a pin can be made
  public from the private view.
- the public edit form passes no isPrivate when re-rendering after an error, so
  PinEditPage falls back to `isPrivate ?? pin.isPrivate`. The private one passes
  the submitted value. The factory passes `undefined` in public mode to keep
  that fallback intact -- passing `false` would have looked equivalent and was
  not.

Also folded up three repetitions inside the merged file: the not-found/not-yours
404 mapping (8 copies), the `?view=` strip-and-carry dance (4 copies), and the
list view props.

All 71 pin-route characterization tests pass unchanged, which is the evidence
this is behavior-preserving. Mutation testing found one gap while checking:
changing the private create flash to the public wording was caught by nothing,
so it now has an assertion -- it is an option rather than a hardcoded string
now, and easy to unify by accident.

The two requirePrivateUnlock registrations are kept as-is. The `/pins` one is
redundant on Hono 4.13, but the cost of being wrong is every private pin
readable while locked, so it stays until something forces the question.
