+++ apps/hono/src/routes/pins.tsx
@@ -1,739 +1,10 @@
 import { Hono } from 'hono'
-import type { Context } from 'hono'
-import {
-  AccessControl,
-  type PinFilter,
-  type User,
-  ValidationError,
-  DuplicatePinError,
-  PinNotFoundError,
-  UnauthorizedPinAccessError,
-} from '@pinsquirrel/domain'
-import { pinService, tagService } from '../lib/services'
-import { parsePinForm } from '../lib/form'
-import {
-  getAuthUser,
-  getSessionManager,
-  requireAuth,
-} from '../middleware/session'
-import { PinCard, PinDeleteConfirm } from '../views/components/PinCard'
-import { PinForm } from '../views/components/PinForm'
-import { PinDeletePage } from '../views/pages/pin-delete'
-import { PinEditPage } from '../views/pages/pin-edit'
-import { PinNewPage } from '../views/pages/pin-new'
-import { PinsPage } from '../views/pages/pins'
-import { PinsContentPartial } from '../views/partials/pins-content'
+import { requireAuth } from '../middleware/session'
+import { createPinRoutes } from './pin-routes'
 
 const pins = new Hono()
 
-// Apply auth middleware to all pin routes
 pins.use('*', requireAuth())
-
-// Helper to parse pin query parameters
-export function parsePinQueryParams(c: Context) {
-  const url = new URL(c.req.url)
-  const tag = url.searchParams.get('tag') || undefined
-  const search = url.searchParams.get('search') || undefined
-  const unreadParam = url.searchParams.get('unread')
-  const notagsParam = url.searchParams.get('notags')
-  const pageParam = url.searchParams.get('page')
-  const sizeParam = url.searchParams.get('size')
-  const sortParam = url.searchParams.get('sort')
-  const directionParam = url.searchParams.get('direction')
-
-  // Build filter — exclude private pins from normal view
-  const filter: PinFilter = { isPrivate: false }
-
-  if (tag) {
-    filter.tag = tag
-  }
-
-  if (search) {
-    filter.search = search
-  }
-
-  // Handle notags filter
-  const noTags = notagsParam === 'true'
-  if (noTags) {
-    filter.noTags = true
-  }
-
-  // Handle unread filter
-  let readFilter: 'all' | 'unread' | 'read' = 'all'
-  if (unreadParam === 'true') {
-    filter.readLater = true
-    readFilter = 'unread'
-  } else if (unreadParam === 'false') {
-    filter.readLater = false
-    readFilter = 'read'
-  }
-
-  // Handle sort settings
-  const sortBy: 'created' | 'title' =
-    sortParam === 'title' ? 'title' : 'created'
-  const sortDirection: 'asc' | 'desc' =
-    directionParam === 'asc' ? 'asc' : 'desc'
-
-  filter.sortBy = sortBy
-  filter.sortDirection = sortDirection
-
-  // Parse page number
-  const page = pageParam ? parseInt(pageParam, 10) : 1
-
-  // Parse view size
-  const viewSize: 'expanded' | 'compact' =
-    sizeParam === 'compact' ? 'compact' : 'expanded'
-
-  // Build search params string for preserving filters (exclude page for base params)
-  const searchParams = url.search.replace(/^\?/, '')
-
-  return {
-    tag,
-    search,
-    readFilter,
-    filter,
-    page,
-    viewSize,
-    sortBy,
-    sortDirection,
-    searchParams,
-    noTags,
-  }
-}
-
-// Helper to fetch pins for a user
-export async function fetchUserPins(
-  user: User,
-  filter: PinFilter,
-  page: number
-) {
-  const ac = new AccessControl(user)
-  return pinService.getUserPinsWithPagination(ac, filter, {
-    page,
-    pageSize: 25,
-  })
-}
-
-// GET /pins - List all pins with filtering and pagination
-pins.get('/', async (c) => {
-  const sessionManager = getSessionManager(c)
-  const user = getAuthUser(c)
-  const isHtmx = !!c.req.header('HX-Request')
-
-  const {
-    tag,
-    search,
-    readFilter,
-    filter,
-    page,
-    viewSize,
-    sortBy,
-    sortDirection,
-    searchParams,
-    noTags,
-  } = parsePinQueryParams(c)
-
-  const result = await fetchUserPins(user, filter, page)
-
-  if (isHtmx) {
-    return c.html(
-      <PinsContentPartial
-        pins={result.pins}
-        pagination={result.pagination}
-        searchParams={searchParams}
-        activeTag={tag}
-        searchQuery={search}
-        readFilter={readFilter}
-        viewSize={viewSize}
-        sortBy={sortBy}
-        sortDirection={sortDirection}
-        noTags={noTags}
-        baseUrl="/pins"
-      />
-    )
-  }
-
-  const flash = sessionManager.getFlash()
-
-  return c.html(
-    <PinsPage
-      user={user}
-      pins={result.pins}
-      pagination={result.pagination}
-      searchParams={searchParams}
-      activeTag={tag}
-      searchQuery={search}
-      readFilter={readFilter}
-      viewSize={viewSize}
-      sortBy={sortBy}
-      sortDirection={sortDirection}
-      noTags={noTags}
-      flash={flash}
-      baseUrl="/pins"
-    />
-  )
-})
-
-// GET /pins/new - Show pin creation form
-pins.get('/new', async (c) => {
-  const sessionManager = getSessionManager(c)
-  const user = getAuthUser(c)
-
-  const ac = new AccessControl(user)
-
-  // Get URL params for bookmarklet integration
-  const url = new URL(c.req.url)
-  const prefillUrl = url.searchParams.get('url') || ''
-  const prefillTitle = url.searchParams.get('title') || ''
-  const prefillDescription = url.searchParams.get('description') || ''
-  const prefillTag = url.searchParams.get('tag') || ''
-
-  // Check if URL already exists for this user (bookmarklet redirect to edit)
-  if (prefillUrl) {
-    const existingPins = await pinService.getUserPinsWithPagination(
-      ac,
-      { url: prefillUrl },
-      { page: 1, pageSize: 1 }
-    )
-
-    if (existingPins.pins.length > 0) {
-      // URL already exists, redirect to edit form
-      return c.redirect(`/pins/${existingPins.pins[0].id}/edit`)
-    }
-  }
-
-  // Fetch user's existing tags for display
-  const userTags = await tagService.getUserTags(ac, user.id)
-
-  // Get flash message if any
-  const flash = sessionManager.getFlash()
-
-  return c.html(
-    <PinNewPage
-      user={user}
-      flash={flash}
-      userTags={userTags.map((t) => t.name)}
-      url={prefillUrl}
-      title={prefillTitle}
-      description={prefillDescription}
-      tags={prefillTag}
-    />
-  )
-})
-
-// POST /pins/new - Create a new pin
-pins.post('/new', async (c) => {
-  const sessionManager = getSessionManager(c)
-  const user = getAuthUser(c)
-
-  const ac = new AccessControl(user)
-
-  // Parse form data
-  const formData = await c.req.parseBody()
-
-  const {
-    url: pinUrl,
-    title,
-    description,
-    readLater,
-    isPrivate,
-    tagsInput,
-    tagNames,
-  } = parsePinForm(formData)
-
-  // Fetch user's tags for re-rendering form on error
-  const userTags = await tagService.getUserTags(ac, user.id)
-
-  try {
-    // Create the pin using service
-    await pinService.createPin(ac, {
-      userId: user.id,
-      url: pinUrl,
-      title,
-      description,
-      readLater,
-      isPrivate,
-      tagNames,
-    })
-
-    // Redirect to pins list with success message
-    sessionManager.setFlash('success', 'Pin created successfully!')
-    if (c.req.header('HX-Request')) {
-      c.header('HX-Redirect', '/pins')
-      return c.body(null)
-    }
-    return c.redirect('/pins')
-  } catch (error) {
-    const isHtmx = !!c.req.header('HX-Request')
-    const userTagNames = userTags.map((t) => t.name)
-
-    const formProps = {
-      action: '/pins/new' as const,
-      submitLabel: 'Create Pin' as const,
-      url: pinUrl,
-      title,
-      description: description || '',
-      readLater,
-      tags: tagsInput,
-      userTags: userTagNames,
-    }
-
-    if (error instanceof ValidationError) {
-      if (isHtmx) {
-        return c.html(<PinForm {...formProps} errors={error.fields} />)
-      }
-      return c.html(
-        <PinNewPage
-          user={user}
-          errors={error.fields}
-          userTags={userTagNames}
-          url={pinUrl}
-          title={title}
-          description={description || ''}
-          readLater={readLater}
-          tags={tagsInput}
-        />
-      )
-    }
-
-    if (error instanceof DuplicatePinError) {
-      const duplicatePinId = error.existingPin?.id
-      if (isHtmx) {
-        return c.html(
-          <PinForm {...formProps} duplicatePinId={duplicatePinId} />
-        )
-      }
-      return c.html(
-        <PinNewPage
-          user={user}
-          duplicatePinId={duplicatePinId}
-          userTags={userTagNames}
-          url={pinUrl}
-          title={title}
-          description={description || ''}
-          readLater={readLater}
-          tags={tagsInput}
-        />
-      )
-    }
-
-    // Generic error
-    const errors = { _form: ['Failed to create pin. Please try again.'] }
-    if (isHtmx) {
-      return c.html(<PinForm {...formProps} errors={errors} />, 500)
-    }
-    return c.html(
-      <PinNewPage
-        user={user}
-        errors={errors}
-        userTags={userTagNames}
-        url={pinUrl}
-        title={title}
-        description={description || ''}
-        readLater={readLater}
-        tags={tagsInput}
-      />,
-      500
-    )
-  }
-})
-
-// GET /pins/:id/edit - Show pin edit form
-pins.get('/:id/edit', async (c) => {
-  const sessionManager = getSessionManager(c)
-  const user = getAuthUser(c)
-
-  const pinId = c.req.param('id')
-  const ac = new AccessControl(user)
-
-  const url = new URL(c.req.url)
-  const returnParams = url.search.replace(/^\?/, '')
-
-  try {
-    // Fetch the pin and user's tags
-    const [pin, userTags] = await Promise.all([
-      pinService.getPin(ac, pinId),
-      tagService.getUserTags(ac, user.id),
-    ])
-
-    // Get flash message if any
-    const flash = sessionManager.getFlash()
-
-    return c.html(
-      <PinEditPage
-        user={user}
-        pin={pin}
-        flash={flash}
-        userTags={userTags.map((t) => t.name)}
-        returnParams={returnParams}
-      />
-    )
-  } catch (error) {
-    if (
-      error instanceof PinNotFoundError ||
-      error instanceof UnauthorizedPinAccessError
-    ) {
-      return c.text('Pin not found', 404)
-    }
-    throw error
-  }
-})
-
-// POST /pins/:id/edit - Update a pin
-pins.post('/:id/edit', async (c) => {
-  const sessionManager = getSessionManager(c)
-  const user = getAuthUser(c)
-
-  const pinId = c.req.param('id')
-  const ac = new AccessControl(user)
-
-  const requestUrl = new URL(c.req.url)
-  const returnParams = requestUrl.search.replace(/^\?/, '')
-  const redirectTarget = returnParams ? `/pins?${returnParams}` : '/pins'
-  const editAction = `/pins/${pinId}/edit${returnParams ? `?${returnParams}` : ''}`
-
-  // Parse form data
-  const formData = await c.req.parseBody()
-
-  // Helper to safely extract string from form data
-  const {
-    url: pinUrl,
-    title,
-    description,
-    readLater,
-    isPrivate,
-    tagsInput,
-    tagNames,
-  } = parsePinForm(formData)
-
-  // Fetch user's tags for re-rendering form on error
-  const userTags = await tagService.getUserTags(ac, user.id)
-
-  try {
-    // Get existing pin to get userId
-    const existingPin = await pinService.getPin(ac, pinId)
-
-    // Update the pin using service
-    await pinService.updatePin(ac, {
-      id: pinId,
-      userId: existingPin.userId,
-      url: pinUrl,
-      title,
-      description,
-      readLater,
-      isPrivate,
-      tagNames,
-    })
-
-    // Redirect to pins list with success message
-    sessionManager.setFlash('success', 'Pin updated successfully!')
-    if (c.req.header('HX-Request')) {
-      c.header('HX-Redirect', redirectTarget)
-      return c.body(null)
-    }
-    return c.redirect(redirectTarget)
-  } catch (error) {
-    // Get the pin for re-rendering the form (may fail if not found)
-    let pin
-    try {
-      pin = await pinService.getPin(ac, pinId)
-    } catch {
-      return c.text('Pin not found', 404)
-    }
-
-    const isHtmx = !!c.req.header('HX-Request')
-    const userTagNames = userTags.map((t) => t.name)
-
-    const formProps = {
-      action: editAction,
-      submitLabel: 'Update Pin' as const,
-      url: pinUrl,
-      title,
-      description: description || '',
-      readLater,
-      tags: tagsInput,
-      userTags: userTagNames,
-      createdAt: pin.createdAt,
-    }
-
-    if (error instanceof ValidationError) {
-      if (isHtmx) {
-        return c.html(<PinForm {...formProps} errors={error.fields} />)
-      }
-      return c.html(
-        <PinEditPage
-          user={user}
-          pin={pin}
-          errors={error.fields}
-          userTags={userTagNames}
-          url={pinUrl}
-          title={title}
-          description={description || ''}
-          readLater={readLater}
-          tags={tagsInput}
-          returnParams={returnParams}
-        />
-      )
-    }
-
-    if (error instanceof DuplicatePinError) {
-      const duplicatePinId = error.existingPin?.id
-      if (isHtmx) {
-        return c.html(
-          <PinForm {...formProps} duplicatePinId={duplicatePinId} />
-        )
-      }
-      return c.html(
-        <PinEditPage
-          user={user}
-          pin={pin}
-          duplicatePinId={duplicatePinId}
-          userTags={userTagNames}
-          url={pinUrl}
-          title={title}
-          description={description || ''}
-          readLater={readLater}
-          tags={tagsInput}
-          returnParams={returnParams}
-        />
-      )
-    }
-
-    if (
-      error instanceof PinNotFoundError ||
-      error instanceof UnauthorizedPinAccessError
-    ) {
-      return c.text('Pin not found', 404)
-    }
-
-    // Generic error
-    const errors = { _form: ['Failed to update pin. Please try again.'] }
-    if (isHtmx) {
-      return c.html(<PinForm {...formProps} errors={errors} />, 500)
-    }
-    return c.html(
-      <PinEditPage
-        user={user}
-        pin={pin}
-        errors={errors}
-        userTags={userTagNames}
-        url={pinUrl}
-        title={title}
-        description={description || ''}
-        readLater={readLater}
-        tags={tagsInput}
-        returnParams={returnParams}
-      />,
-      500
-    )
-  }
-})
-
-// POST /pins/:id/toggle-read - Toggle read later status (HTMX)
-pins.post('/:id/toggle-read', async (c) => {
-  const user = getAuthUser(c)
-
-  const pinId = c.req.param('id')
-  const ac = new AccessControl(user)
-
-  // Get the current pin to toggle its status
-  const existingPin = await pinService.getPin(ac, pinId)
-
-  // Toggle the readLater status
-  const updatedPin = await pinService.updatePin(ac, {
-    id: existingPin.id,
-    userId: existingPin.userId,
-    url: existingPin.url,
-    title: existingPin.title,
-    description: existingPin.description,
-    readLater: !existingPin.readLater,
-    isPrivate: existingPin.isPrivate,
-    tagNames: existingPin.tagNames,
-  })
-
-  // Get search params from referer or request for preserving filters
-  const referer = c.req.header('Referer') || ''
-  let searchParams = ''
-  try {
-    const refererUrl = new URL(referer)
-    searchParams = refererUrl.search.replace(/^\?/, '')
-  } catch {
-    // Ignore invalid referer
-  }
-
-  // Return updated PinCard for HTMX to swap
-  return c.html(
-    <PinCard
-      pin={updatedPin}
-      viewSize="expanded"
-      searchParams={searchParams}
-      baseUrl="/pins"
-    />
-  )
-})
-
-// GET /pins/:id/delete-confirm - Inline delete confirmation (HTMX)
-pins.get('/:id/delete-confirm', async (c) => {
-  const user = getAuthUser(c)
-
-  const pinId = c.req.param('id')
-  const ac = new AccessControl(user)
-  const url = new URL(c.req.url)
-  const viewSize = (url.searchParams.get('view') || 'expanded') as
-    'expanded' | 'compact'
-  url.searchParams.delete('view')
-  const searchParams = url.searchParams.toString()
-
-  try {
-    const pin = await pinService.getPin(ac, pinId)
-
-    return c.html(
-      <PinDeleteConfirm
-        pin={pin}
-        viewSize={viewSize}
-        searchParams={searchParams}
-        baseUrl="/pins"
-      />
-    )
-  } catch (error) {
-    if (
-      error instanceof PinNotFoundError ||
-      error instanceof UnauthorizedPinAccessError
-    ) {
-      return c.text('Pin not found', 404)
-    }
-    throw error
-  }
-})
-
-// DELETE /pins/:id - Delete a pin (HTMX, returns refreshed list)
-pins.delete('/:id', async (c) => {
-  const user = getAuthUser(c)
-
-  const pinId = c.req.param('id')
-  const ac = new AccessControl(user)
-
-  try {
-    await pinService.deletePin(ac, pinId)
-
-    const {
-      tag,
-      search,
-      readFilter,
-      filter,
-      page,
-      viewSize,
-      sortBy,
-      sortDirection,
-      searchParams,
-      noTags,
-    } = parsePinQueryParams(c)
-    const result = await fetchUserPins(user, filter, page)
-
-    return c.html(
-      <PinsContentPartial
-        pins={result.pins}
-        pagination={result.pagination}
-        searchParams={searchParams}
-        activeTag={tag}
-        searchQuery={search}
-        readFilter={readFilter}
-        viewSize={viewSize}
-        sortBy={sortBy}
-        sortDirection={sortDirection}
-        noTags={noTags}
-        baseUrl="/pins"
-      />
-    )
-  } catch (error) {
-    if (
-      error instanceof PinNotFoundError ||
-      error instanceof UnauthorizedPinAccessError
-    ) {
-      return c.text('Pin not found', 404)
-    }
-    throw error
-  }
-})
-
-// GET /pins/:id/card - Return a single pin card (HTMX, for cancel)
-pins.get('/:id/card', async (c) => {
-  const user = getAuthUser(c)
-
-  const pinId = c.req.param('id')
-  const ac = new AccessControl(user)
-  const url = new URL(c.req.url)
-  const viewSize = (url.searchParams.get('view') || 'expanded') as
-    'expanded' | 'compact'
-  url.searchParams.delete('view')
-  const searchParams = url.searchParams.toString()
-
-  try {
-    const pin = await pinService.getPin(ac, pinId)
-
-    return c.html(
-      <PinCard
-        pin={pin}
-        viewSize={viewSize}
-        searchParams={searchParams}
-        baseUrl="/pins"
-      />
-    )
-  } catch (error) {
-    if (
-      error instanceof PinNotFoundError ||
-      error instanceof UnauthorizedPinAccessError
-    ) {
-      return c.text('Pin not found', 404)
-    }
-    throw error
-  }
-})
-
-// GET /pins/:id/delete - Show delete confirmation (full page, non-JS fallback)
-pins.get('/:id/delete', async (c) => {
-  const user = getAuthUser(c)
-
-  const pinId = c.req.param('id')
-  const ac = new AccessControl(user)
-
-  try {
-    const pin = await pinService.getPin(ac, pinId)
-
-    return c.html(<PinDeletePage user={user} pin={pin} />)
-  } catch (error) {
-    if (
-      error instanceof PinNotFoundError ||
-      error instanceof UnauthorizedPinAccessError
-    ) {
-      return c.text('Pin not found', 404)
-    }
-    throw error
-  }
-})
-
-// POST /pins/:id/delete - Delete a pin
-pins.post('/:id/delete', async (c) => {
-  const sessionManager = getSessionManager(c)
-  const user = getAuthUser(c)
-
-  const pinId = c.req.param('id')
-  const ac = new AccessControl(user)
-
-  try {
-    await pinService.deletePin(ac, pinId)
-
-    sessionManager.setFlash('success', 'Pin deleted successfully!')
-    return c.redirect('/pins')
-  } catch (error) {
-    if (
-      error instanceof PinNotFoundError ||
-      error instanceof UnauthorizedPinAccessError
-    ) {
-      return c.text('Pin not found', 404)
-    }
-    throw error
-  }
-})
+pins.route('/', createPinRoutes({ baseUrl: '/pins' }))
 
 export { pins as pinsRoutes }
